# Frontend/dividend-portfolio-manager/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Questrade Portfolio Tracker API
  description: |
    A comprehensive Node.js application for tracking your Questrade portfolio with automatic token refresh, 
    real-time data sync, and detailed dividend tracking.
    
    ## Features
    - 🔐 Automatic token refresh (weekly)
    - 📊 Real-time portfolio data sync
    - 💰 Detailed dividend tracking and analysis
    - 📈 Portfolio performance metrics
    - 🎯 Snap quote integration for current prices
    - 📅 Historical snapshots
    - 🔄 Automatic data synchronization during market hours
    
    ## Multi-Person Support
    This API supports multiple persons/users, each with their own Questrade account(s) and tokens.
    
    ## Authentication
    Each person requires a valid Questrade refresh token. The system automatically manages access tokens.
    
  version: "2.0.1"
  contact:
    name: Portfolio Manager Support
    url: https://github.com/yourusername/portfolio-manager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
...
paths:
...
  /auth/setup-person:
    post:
      tags: ["Authentication & Tokens"]
      summary: Setup authentication for a new person
      description: Initialize a person with their Questrade refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [personName, refreshToken]
              properties:
                personName:
                  type: string
                  example: "John Doe"
                refreshToken:
                  type: string
                  example: "sBPZVcmNB6udi8zcwT8uqIgm2jAkFdvI0"
      responses:
        '201':
          description: Person setup successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh-token/{personName}:
    put:
      tags: ["Authentication & Tokens"]
      summary: Refresh access token for person
      parameters:
        - $ref: '#/components/parameters/PersonName'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          expiresIn:
                            type: integer
                            example: 1800
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/token-status/{personName}:
    get:
      tags: ["Authentication & Tokens"]
      summary: Get token status for person
      parameters:
        - $ref: '#/components/parameters/PersonName'
      responses:
        '200':
          description: Token status retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenStatus'
...
  /auth/test-connection/{personName}:
    post:
      tags: ["Authentication & Tokens"]
      summary: Test API connection for person
      parameters:
        - $ref: '#/components/parameters/PersonName'
      responses:
        '200':
          description: Connection test successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          serverTime:
                            type: string
                            format: date-time
                          responseTime:
                            type: integer

  # Person Management
  /persons:
    get:
      tags: ["Person Management"]
      summary: Get all persons
      responses:
        '200':
          description: List of all persons
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Person'

    post:
      tags: ["Person Management"]
      summary: Create new person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [personName, refreshToken]
              properties:
                personName:
                  type: string
                  example: "Jane Smith"
                refreshToken:
                  type: string
                  example: "sBPZVcmNB6udi8zcwT8uqIgm2jAkFdvI0"
                displayName:
                  type: string
                  example: "Jane S."
                email:
                  type: string
                  format: email
                  example: "jane@example.com"
                phoneNumber:
                  type: string
                  example: "+1-555-0123"
      responses:
        '201':
          description: Person created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Person'
  /persons/{personName}:
    get:
      tags: ["Person Management"]
      summary: Get specific person details
      parameters:
        - $ref: '#/components/parameters/PersonName'
      responses:
        '200':
          description: Person details retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PersonDetails'

    put:
      tags: ["Person Management"]
      summary: Update person information
      parameters:
        - $ref: '#/components/parameters/PersonName'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newPersonName:
                  type: string
                displayName:
                  type: string
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
                preferences:
                  $ref: '#/components/schemas/PersonPreferences'
      responses:
        '200':
          description: Person updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Person'

    delete:
      tags: ["Person Management"]
      summary: Delete person (soft delete)
      parameters:
        - $ref: '#/components/parameters/PersonName'
        - name: permanent
          in: query
          schema:
            type: boolean
            default: false
          description: Whether to permanently delete all data
      responses:
        '200':
          description: Person deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Accounts
...
  /sync/person/{personName}:
    post:
      tags: ["Data Synchronization"]
      summary: Sync data for specific person
      parameters:
        - $ref: '#/components/parameters/PersonName'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullSync:
                  type: boolean
                  default: false
                  description: Whether to perform full sync (6 months of data)
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SyncResult'
...
components:
  parameters:
    PersonName:
      name: personName
      in: path
      required: true
      schema:
        type: string
      example: "John Doe"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or expired token
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  errorCode:
                    type: string
                    example: "TOKEN_EXPIRED"
                  userMessage:
                    type: string
                    example: "Your access token has expired. The system will automatically refresh it on the next request."
                  recoverySuggestions:
                    type: array
                    items:
                      type: string
                    example: ["The system will automatically refresh your token", "If this persists, check your refresh token in Settings"]
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        timestamp:
          type: string
          format: date-time
    TokenStatus:
      type: object
      properties:
        personName:
          type: string
        refreshToken:
          type: object
          properties:
            exists:
              type: boolean
            expiresAt:
              type: string
              format: date-time
            lastUsed:
              type: string
              format: date-time
            errorCount:
              type: integer
            lastError:
              type: string
        accessToken:
          type: object
          properties:
            exists:
              type: boolean
            expiresAt:
              type: string
              format: date-time
            apiServer:
              type: string
        isHealthy:
          type: boolean
    Person:
      type: object
      properties:
        personName:
          type: string
        displayName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        isActive:
          type: boolean
        hasValidToken:
          type: boolean
        lastTokenRefresh:
          type: string
          format: date-time
        lastSuccessfulSync:
          type: string
          format: date-time
        preferences:
          $ref: '#/components/schemas/PersonPreferences'
        createdAt:
          type: string
          format: date-time
    PersonDetails:
      allOf:
        - $ref: '#/components/schemas/Person'
        - type: object
          properties:
            accounts:
              type: array
              items:
                $ref: '#/components/schemas/Account'
            tokenStatus:
              $ref: '#/components/schemas/TokenStatus'
            statistics:
              type: object
              properties:
                positionCount:
                  type: integer
                activityCount:
                  type: integer
                accountCount:
                  type: integer
    PersonPreferences:
      type: object
      properties:
        defaultView:
          type: string
          enum: [all, person, account]
        currency:
          type: string
          enum: [CAD, USD]
        notifications:
          type: object
          properties:
            enabled:
              type: boolean
            dividendAlerts:
              type: boolean
            syncErrors:
              type: boolean
    Account:
      type: object
      properties:
        accountId:
          type: string
        personName:
          type: string
        type:
          type: string
        status:
          type: string
        isPrimary:
          type: boolean
        balances:
          $ref: '#/components/schemas/AccountBalances'
        numberOfPositions:
          type: integer
        totalInvestment:
          type: number
        currentValue:
          type: number
        dayPnl:
          type: number
        openPnl:
          type: number
        closedPnl:
          type: number
        totalPnl:
          type: number
        syncedAt:
          type: string
          format: date-time
    AccountBalances:
      type: object
      properties:
        perCurrencyBalances:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyBalance'
        combinedBalances:
          $ref: '#/components/schemas/CurrencyBalance'
        lastUpdated:
          type: string
          format: date-time
    CurrencyBalance:
      type: object
      properties:
        currency:
          type: string
        cash:
          type: number
        marketValue:
          type: number
        totalEquity:
          type: number
        buyingPower:
          type: number
        maintenanceExcess:
          type: number
        isRealTime:
          type: boolean
...
    SyncResult:
      type: object
      properties:
        personName:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
        accounts:
          type: object
          properties:
            synced:
              type: integer
            errors:
              type: array
              items:
                $ref: '#/components/schemas/SyncError'
        positions:
          type: object
          properties:
            synced:
              type: integer
            errors:
              type: array
              items:
                $ref: '#/components/schemas/SyncError'
        activities:
          type: object
          properties:
            synced:
              type: integer
            errors:
              type: array
              items:
                $ref: '#/components/schemas/SyncError'
        snapshots:
          type: object
          properties:
            created:
              type: boolean
            error:
              type: string
    SyncError:
      type: object
      properties:
        accountId:
          type: string
        symbol:
          type: string
        type:
          type: string
        error:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)

# Security (if you implement API key authentication)
# security:
#   - ApiKeyAuth: []